+++ zk/tox.ini
@@ -10,7 +10,6 @@ platform = linux|darwin|win32
 deps =
     ../datadog_checks_base
     -r../datadog_checks_base/requirements.in
-    ../datadog_checks_tests_helper
     -rrequirements-dev.txt
 commands =
     pip install --require-hashes -r requirements.txt
